tests.integration: remove also record identifiers
* Also refactored the `_delete_record` and reused on the tests.
